Reorganize vector mode tests into proper unit test structure with type hints
